+++ documentacion/Casos.txt
@@ -1,0 +1,18 @@
+Faltan Agregar:
+
+- Propina
+- Cashback 
+
+-- Transaccion Online Propina
+Select TOP 2 Numtar,NroAut, Importe , Cuotas from StratusMovims02Diario WITH(NOLOCK)
+where  NroProc = @NumeroProceso and Cuotas = 999 and mk4 = 'T'
+
+
+
+-- Transacción Cashback - Emisión Prisma y MOV2000
+Select TOP 2 Numtar,NroAut, Importe , Cuotas from StratusMovims02Diario WITH(NOLOCK)
+where  NroProc = @NumeroProceso and Cuotas = 998 and mk4 = 'T'
+
+-- Transacción Cashback - Emisión NO Prisma y MOV2000
+Select TOP 2 Numtar,NroAut, Importe , Cuotas from StratusMovims02Diario WITH(NOLOCK)
+where  NroProc = @NumeroProceso and Cuotas = 998 and TjCodBco = 998
