+++ documentacion/TARJETA_CONFIABLE011512_00003.txt
@@ -1,0 +1,26 @@
+H2020011509270800003                                                                                                                                                                    
+D100000456428500120001895607180052528359   02202019121014283100000040032000000044400010615880000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121014284400000041032000000050333012017290000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121014284800000042032000000055500122086160000000                         00050       0001201912107000                                
+D000000456428500120001895607180052528359   02202019121014295500000043032000000055500121527860000000                         00050       0001201912107000                                
+D000000456428500120001895607180052528359   02202019121015062200000047032000000044400017201250000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015062400000048032000000044400012349860000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015062500000049032000000044400015776510000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015063100000050032000000050333016843700000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015063900000051032000000055500129249600000000                         00050       0001201912107000                                
+D000000456428500120001895607180052528359   02202019121015063900000052032000000055500122022450000000                         00050       0001201912107000                                
+D000000456428500120001895607180052528359   02202019121015092700000053032000000044400014176400000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015092900000054032000000044400011736280000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015093100000055032000000044400015631590000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015093200000056032000000044400010245110000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015093700000057032000000050333018044650000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015094000000058032000000050333015054660000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015094100000059032000000050333010992540000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   02202019121015094600000060032000000055500126902600000000                         00050       0001201912107000                                
+D000000456428500120001895607180052528359   02202019121015094700000061032000000055500129082580000000                         00050       0001201912107000                                
+D000000456428500120001895607180052528359   02202019121015094800000062032000000055500125572190000000                         00050       0001201912107000                                
+D000000456428500120001895607180052528359   02202019121015293100000065032000000050333018697880000000                         00050       0001201912100000                                
+D000000456428500120001895607180052528359   03202019121017462300000072032000000050333015963830000000                         60000       0001201912100000                                
+D000000456428500120001895607180052528359   03202019121017414200000071032000000055500123988960000000                         60000       0001201912100000                                
+T000000456428500120001890000002100000104569800000000000000000000000000020000001058330000000000000000000000000000                                                                        
+G0000002100000104569800000000000000000000000000020000001058330000000000000000000000000000
