+++ documentacion/para el exe.txt
@@ -1,0 +1,1 @@
+pyinstaller --onefile --icon=validator_icono.ico main.pyw
