+++ setUp.pyw
@@ -1,21 +1,17 @@
-#-----VALIDATOR-------#
+ # #Importar las librerias
 import tkinter as tk
-from tkinter import filedialog
+from tkinter import filedialog, ttk, messagebox
+from tkinter.ttk import *
 from tkinter import *
-from tkinter import ttk
+import time as time
 import os
-import Grilla
+#import Grilla
 import wx
 import wx.grid as gridlib
 from timeit import timeit
-import time as time
-# IMPORTAR LOS ARCHIVOS
-import tkinter as tk
-import archivos
-from tkinter import *
-from tkinter.ttk import *
-import time
-from tkinter import messagebox
+#import archivos
+from vicap import *
+from vicap import archivos, Grilla, archivoMkpriv, GrillaMkpriv
 
 class OpenForms(): 
 
@@ -30,15 +26,14 @@ class OpenForms():
 		
 		#METODO QUE OBTIENE LA RUTA DONDE ESTA EL ARCHIVO MOV2000
 		def abrir_archivo():
-			archivo_abierto=filedialog.askopenfilename(initialdir = "/",
+			archivo_abierto=filedialog.askopenfilename(initialdir = "/", #esto abre en el raiz. ver de mejorar y poner donde esta el proyecto
                 title = "Seleccione archivo",filetypes = (("txt files","*.txt"),
                 ("all files","*.*")))
 			return(archivo_abierto)
 
-		#BOTÓN LEER QUE LLAMA AL MÉTODO CLICKED
-		def clicked():
-	
-					
+		
+		def clickedTotal(): #casos totales
+						
 			mov2000Plano = "a"
 			listaArchivo = []
 			contadorArchivo = 0
@@ -47,7 +42,6 @@ class OpenForms():
 			ruta= ""
 
 			ruta = abrir_archivo() #SE LLAMA EL MÉTODO DONDE SE CONSIGUE LA RUTA
-
 
 			if ruta != '':  #Si la ruta no está vacía proceso.
 	
@@ -90,11 +84,9 @@ class OpenForms():
 					print("Formato de Archivo erroneo, vuelvo a iterar")
 					messagebox.showinfo(message="¡Error en memoria, archivo muy grande!", title="Error")
 			else:
-				print("Vuelvo a iterar")
-		
-		
+				print("Vuelvo a iterar")	#esto es cuando se abre y se cierra la ventana con cancelar				
 
-		def clicked2():
+		def clickedParticular(): #casos particulares
 	
 					
 			mov2000Plano = "a"
@@ -131,22 +123,76 @@ class OpenForms():
 						OpenForms.abrirFormulario()
 
 					else:
-						messagebox.showinfo(message="¡Error en formato de Archivo!", title="Error")
+						messagebox.showinfo(message="¡Error en formato de Archivo! AAA", title="Error")
 						
 						print("Formato de Archivo erroneo, vuelvo a iterar")
 				else:
 					print("Formato de Archivo erroneo, vuelvo a iterar")
-					messagebox.showinfo(message="¡Error en formato de Archivo!", title="Error")
+					messagebox.showinfo(message="¡Error en formato de Archivo! BBB", title="Error")
 
 			else:
 				print("Vuelvo a iterar")
 
+		def clickedMkpriv(): #casos marcas privadas
+						
+			mkprivPlano = "a"
+			listaArchivo = []
+			contadorArchivo = 0
+			CSV_MKPRIV = "a"
+			listaCompleta = []
+			ruta= ""
 
+			ruta = abrir_archivo() #SE LLAMA EL MÉTODO DONDE SE CONSIGUE LA RUTA
+
+			if ruta != '':  #Si la ruta no está vacía proceso.
+	
+				mkprivPlano = archivoMkpriv.manejoDeLosArchivosTXT.abrirArchivo(mkprivPlano,ruta)
+				listaArchivo, contadorArchivo = archivoMkpriv.manejoDeLosArchivosTXT.recorrerArchivoMkpriv(mkprivPlano, listaArchivo, contadorArchivo)
+
+				if listaArchivo !=11: #Si el método listaArchivo no dió excepción de memoria entro
+
+					if listaArchivo != 10: #Si el método listaArchivo no dió otra excepcion entro
+						listaCompleta = archivoMkpriv.manejoDeLosArchivosTXT.subStringLista(listaArchivo, contadorArchivo, listaCompleta)
+						archivoMkpriv.manejoDeLosArchivosTXT.cerrarArchivo(mkprivPlano)
+
+						if listaCompleta != 10: #Si el método listaCompleta no dió excepcion entro
+							
+							if len(listaArchivo) < 100000: #Si el MOV2000 es muy grande no carga la grilla
+								window.destroy()
+								import ventanaProgress	
+								app = wx.App()
+								print("Tamaño Lista: ", len(listaArchivo))
+								ventana = ventanaProgress.start1()
+								display = GrillaMkpriv.MyForm().Show()
+								ventana.destroy()						
+								app.MainLoop()
+								timestamp = time.strftime('%Y%m%d%H%M%S')
+								os.rename('CSV_MKPRIV.CSV', 'TOTALMKPRIV_'+timestamp+'.CSV')
+								OpenForms.abrirFormulario()
+							else:
+								messagebox.showinfo(message="¡Archivo muy grande para mostrar en grilla!", title="Error")
+								print("Formato de muy grande vuelvo a iterar")
+
+
+						else:
+							messagebox.showinfo(message="¡Error en formato de Archivo!", title="Error")
+							print("Formato de Archivo erroneo, vuelvo a iterar")
+					else:
+						print("Formato de Archivo erroneo, vuelvo a iterar")
+						messagebox.showinfo(message="¡Error en formato de Archivo!", title="Error")
+
+				else:
+					print("Formato de Archivo erroneo, vuelvo a iterar")
+					messagebox.showinfo(message="¡Error en memoria, archivo muy grande!", title="Error")
+			else:
+				print("Vuelvo a iterar")	#esto es cuando se abre y se cierra la ventana con cancelar				
 			
-		btn = Button(window, text="Casos Particulares", command=clicked2)
+		btn = Button(window, text="Casos Particulares", command=clickedParticular)
 		btn.pack(expand= "True",fill="x")
-		btn2 = Button(window, text="Lectura Total", command=clicked)
+		btn2 = Button(window, text="Lectura Total", command=clickedTotal)
 		btn2.pack(expand= "True",fill="x")
+		btn3 = Button(window, text="Marcas Privadas", command=clickedMkpriv)
+		btn3.pack(expand= "True",fill="x")
 
 	
 		window.mainloop()
